Remove Var.dvals at the top of the call stack [fixes #4]
If any code causes Var.dvals to be set outside of a push/pop pair, and a
later pop never occurs to remove it, then it will cause the thread that
called .invoke to hold a reference to Var.dvals, which will cause all of
the classes loaded in the pod to be retained.

This removes dvals, but only if we are the outer .invoke on the current
thread (in case code invoked reaches out of the shim then back in).